Extract readTrie and parseStateCommitment funcs
